+++ cli/src/main/java/denominator/cli/ResourceRecordSetCommands.java
@@ -20,11 +20,14 @@ import java.net.URI;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import com.google.common.base.Function;
 import com.google.common.base.Joiner;
 import com.google.common.base.Splitter;
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableMap;
 
 import denominator.DNSApiManager;
 import denominator.ResourceRecordSetApi;
@@ -43,6 +46,7 @@ import denominator.model.rdata.SOAData;
 import denominator.model.rdata.SPFData;
 import denominator.model.rdata.SRVData;
 import denominator.model.rdata.TXTData;
+import denominator.route53.AliasTarget;
 
 class ResourceRecordSetCommands {
 
@@ -140,6 +144,16 @@ class ResourceRecordSetCommands {
     }
 
     private static abstract class ModifyRecordSetCommand extends ResourceRecordSetCommand {
+        static final Pattern ELB_REGION = Pattern.compile("[^.]+\\.([^.]+)\\.elb\\.amazonaws\\.com\\.?");
+
+        static final Map<String, String> REGION_TO_HOSTEDZONE = ImmutableMap.<String, String> builder()//
+                .put("us-east-1", "Z3DZXE0Q79N41H")//
+                .put("us-west-2", "Z33MTJ483KN6FU")//
+                .put("eu-west-1", "Z3NF1Z3NOM5OY2")//
+                .put("ap-northeast-1", "Z2YN17T5R711GT")//
+                .put("ap-southeast-1", "Z1WI8VXHPB1R38")//
+                .put("sa-east-1", "Z2ES78Y61JGQKS").build();
+
         @Option(type = OptionType.COMMAND, required = true, name = { "-n", "--name" }, description = "name of the record set. ex. www.denominator.io.")
         public String name;
 
@@ -160,6 +174,9 @@ class ResourceRecordSetCommands {
 
         @Option(type = OptionType.COMMAND, required = false, name = "--ec2-local-hostname", description = "take data from EC2 Instance Metadata local-hostname")
         public boolean ec2LocalHostname;
+
+        @Option(type = OptionType.COMMAND, required = false, name = "--elb-dnsname", description = "dnsname of the ELB to alias. ex. nccp-cbp-frontend-12345678.us-west-2.elb.amazonaws.com.")
+        public String elbDNSName;
 
         public URI metadataService = InstanceMetadataHook.DEFAULT_URI;
 
@@ -182,10 +199,19 @@ class ResourceRecordSetCommands {
                 addIfPresentInMetadataService(valuesBuilder, "local-hostname", metadataService);
             }
             values = valuesBuilder.build();
-            checkArgument(values.size() > 0, "you must pass data to add");
+            checkArgument(elbDNSName != null || values.size() > 0, "you must pass data to add");
             Builder<Map<String, Object>> builder = ResourceRecordSet.builder().name(name).type(type);
-            for (String value : values) {
-                builder.add(toMap(type, value));
+            if (elbDNSName != null) {
+                Matcher getRegion = ELB_REGION.matcher(elbDNSName);
+                checkArgument(getRegion.matches(), "expected elb %s to match %s", ELB_REGION, elbDNSName);
+                String hostedZoneId = REGION_TO_HOSTEDZONE.get(getRegion.group(1));
+                checkArgument(hostedZoneId != null, "region %s not in configured regions: %s", getRegion.group(1),
+                        REGION_TO_HOSTEDZONE.keySet());
+                builder.add(AliasTarget.create(hostedZoneId, elbDNSName));
+            } else {
+                for (String value : values) {
+                    builder.add(toMap(type, value));
+                }
             }
             return builder;
         }